@@ -11,31 +11,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.0.5",
-    "@angular/common": "^11.0.5",
-    "@angular/compiler": "^11.0.5",
-    "@angular/core": "^11.0.5",
-    "@angular/forms": "^11.0.5",
-    "@angular/platform-browser": "^11.0.5",
-    "@angular/platform-browser-dynamic": "^11.0.5",
-    "@angular/router": "^11.0.5",
+    "@angular/animations": "^12.0.4",
+    "@angular/common": "^12.0.4",
+    "@angular/compiler": "^12.0.4",
+    "@angular/core": "^12.0.4",
+    "@angular/forms": "^12.0.4",
+    "@angular/platform-browser": "^12.0.4",
+    "@angular/platform-browser-dynamic": "^12.0.4",
+    "@angular/router": "^12.0.4",
     "bootstrap": "^3.4.1",
+    "install": "^0.13.0",
     "rxjs": "^6.6.3",
     "rxjs-compat": "^6.0.0",
     "tslib": "^2.0.3",
-    "zone.js": "^0.10.3"
+    "update": "^0.4.2",
+    "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.4",
-    "@angular/cli": "^12.0.3",
-    "@angular/compiler-cli": "^11.0.5",
+    "@angular-devkit/build-angular": "^12.0.4",
+    "@angular/cli": "^12.0.4",
+    "@angular/compiler-cli": "^12.0.4",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.19.9",
     "codelyzer": "^6.0.1",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.0",
+    "karma": "^6.3.3",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
@@ -43,6 +45,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "typescript": "~4.2.4"
   }
 }
