@@ -1,5 +1,12 @@
-import { Component, OnInit, Output, EventEmitter } from "@angular/core";
+import {
+  Component,
+  OnInit,
+  Output,
+  EventEmitter,
+  OnDestroy,
+} from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
+import { Subscription } from "rxjs";
 
 import { Recipe } from "../recipe.model";
 import { RecipeService } from "../recipe.service";
@@ -9,10 +16,11 @@ import { RecipeService } from "../recipe.service";
   templateUrl: "./recipe-list.component.html",
   styleUrls: ["./recipe-list.component.css"],
 })
-export class RecipeListComponent implements OnInit {
+export class RecipeListComponent implements OnInit, OnDestroy {
   recipes: Recipe[] = [];
   @Output() recipeClickRedirect = new EventEmitter<Recipe>();
   recipeDetails: Recipe;
+  sub: Subscription;
   navRecipeDetails(recipe: Recipe) {
     // console.log("click resived");
     // console.log(recipe);
@@ -30,6 +38,14 @@ export class RecipeListComponent implements OnInit {
   }
 
   ngOnInit() {
+    this.sub = this.recipeServise.recipesChanged.subscribe(
+      (recipes: Recipe[]) => {
+        this.recipes = recipes;
+      }
+    );
     this.recipes = this.recipeServise.getRecipes();
+  }
+  ngOnDestroy() {
+    this.sub.unsubscribe();
   }
 }
